refactor: migrate のロジックを internal/app/migrate へ移設し main.go を薄くする
- cmd/migrate/main.go を batch と同じくロガー設定と os.Exit のみの薄いラッパーに変更
- allowedCommands・ディスパッチ・goose 実行ロジックを internal/app/migrate へ移設
- cmd/migrate/main_test.go を internal/app/migrate/migrate_test.go へ移設（run → Run）
- cmd/batch/main.go にも JSON ロガー設定を追加し、2バイナリの出力を統一
- README.md / CLAUDE.md のディレクトリ構成に app/batch・config・migrate を追記
