+++ README.md
@@ -82,7 +82,10 @@ REST APIとして、ユーザー認証・株式データ配信・キャッシュ
 │   │   └── types.gen.go        # 生成コード（手動編集不可）
 │   │
 │   ├── app/                    # アプリケーション基盤
+│   │   ├── batch/              # バッチ実行ロジック（job_id ディスパッチ: candles / logo）
+│   │   ├── config/             # 環境変数パースの純粋関数ヘルパー
 │   │   ├── di/                 # 依存性注入
+│   │   ├── migrate/            # マイグレーション実行ロジック（goose サブコマンドディスパッチ）
 │   │   └── router/             # ルーティング設定
 │   │
 │   ├── feature/                # フィーチャーモジュール（垂直スライス）
